[refactor] extract optional OpenCV operation provider
Main changes:
- move OpenCV-backed CPU operations behind a configurable private provider
- keep core registration dependency-neutral and compose configured providers in PluginManager

Files:
- src/lib/providers/opencv/: own OpenCV algorithms, process policy, exception fences, and test access
- src/lib/core/ops.cpp: retain only dependency-neutral core operations
- src/lib/providers/configured_operation_providers.cpp: compose core and enabled provider registrations
- CMakeLists.txt: add the default-on provider option and conditional targets
- tests/: follow the provider-owned access boundary and removed legacy alias

Traceability:
- Motivation: establish the Issue #58 optional operation-provider boundary
- Impact: OpenCV operations remain enabled by default; disabling the provider removes those built-ins without changing the public ABI
- Tests: focused targets built; related CTest selection passed 67/67
- Risk: medium because operation registration ownership and exception boundaries moved
