+++ src/lib/core/ops.hpp
@@ -25,18 +25,20 @@ int builtin_input_halo_radius(const std::string& type,
                               const plugin::ParameterMap& parameters) noexcept;
 
 /**
- * @brief Initializes the OpenCV CPU policy and publishes built-in operations.
+ * @brief Publishes dependency-neutral built-in operations.
+ *
+ * The core registration set currently contains named-data analyzers and scalar
+ * math callbacks plus their propagation contracts. Optional image algorithm
+ * providers are composed separately by the process plugin owner.
+ *
  * @return Nothing.
  * @throws std::bad_alloc if registry key or callback storage allocation fails.
- * @throws std::system_error if one-time OpenCV initialization cannot
- *         synchronize.
- * @note The process plugin owner calls this before built-in callbacks become
- *       visible. `cv::setNumThreads(1)` runs exactly once to prevent nested
- *       OpenCV CPU parallelism; repository code never reconfigures it while
- *       callbacks may execute. Registered CPU providers use `cv::Mat`, own no
- *       shared operation mutex, and rely on scheduler worker admission.
+ * @note This function has no OpenCV or provider dependency. Repeated calls
+ *       replace the same core slots and are used by deterministic restoration
+ *       tests; normal process startup calls it once through the configured
+ *       provider composition boundary.
  */
-void register_builtin();
+void register_core_operations();
 
 }  // namespace ops
 }  // namespace ps
